feat: add common utils for webpack and vite plugin
