+++ packages/utils/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "tailwindcss-webpack-plugin-utils",
+  "version": "0.0.1",
+  "description": "common utils for tailwindcss-webpack-plugin and tailwindcss-vite-plugin",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist/*"
+  ],
+  "scripts": {
+    "build": "tsup",
+    "dev": "pnpm run build -- --watch"
+  },
+  "dependencies": {
+    "fast-glob": "^3.2.12",
+    "normalize-path": "^3.0.0",
+    "debug": "^4.3.4",
+    "postcss": "^8.4.17"
+  },
+  "devDependencies": {
+    "@types/normalize-path": "^3.0.0",
+    "@types/debug": "^4.1.7",
+    "@types/tailwindcss": "^3.1.0",
+    "tsup": "^6.2.3",
+    "tailwindcss": "^3.1.8"
+  },
+  "peerDependencies": {
+    "tailwindcss": "^3.1.8"
+  }
+}
