+++ packages/utils/src/constants.ts
@@ -1,0 +1,4 @@
+export const DEV_TOOLS_VIRTUAL_ID = '_tailwind-devtools_.js';
+export const TAILWIND_ENTRY_VIRTUAL_ID = '_tailwind_.css';
+export const DEFAULT_TAILWIND_ENTRY_CONTENT = `@tailwind base;\n@tailwind components;\n@tailwind utilities;`;
+export const DEFAULT_TAILWIND_CONFIG_FILE = 'tailwind.config.js';
