+++ packages/utils/test/fixtures/load-config/dep-a.js
