+++ packages/utils/test/fixtures/load-config/dep-b.js
