+++ packages/utils/test/fixtures/load-config/dep-c.js
