+++ packages/utils/test/fixtures/scan-contents/tailwind.config.js
