+++ packages/webpack-plugin/package.json
@@ -2,7 +2,7 @@
   "version": "0.0.7",
   "name": "tailwindcss-webpack-plugin",
   "license": "MIT",
-  "description": "Design in Devtools for Tailwind CSS in webpack",
+  "description": "Design in Devtools for Tailwind CSS in Webpack",
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
   "files": [
@@ -27,22 +27,19 @@
   },
   "scripts": {
     "build": "tsup",
-    "dev": " npm run build -- --watch"
+    "dev": "pnpm run build -- --watch"
   },
   "dependencies": {
-    "@types/normalize-path": "^3.0.0",
-    "debug": "^4.3.4",
-    "fast-glob": "^3.2.12",
     "get-port-please": "^2.6.1",
-    "normalize-path": "^3.0.0",
-    "postcss": "^8.4.17",
     "tailwind-config-viewer": "^1.7.2",
-    "tailwindcss": "^3.1.8"
+    "tailwindcss": "^3.1.8",
+    "tailwindcss-webpack-plugin-utils": "workspace:*"
   },
   "peerDependencies": {
     "webpack": "^5.72.0"
   },
   "devDependencies": {
+    "@types/normalize-path": "^3.0.0",
     "@types/debug": "^4.1.7",
     "@types/tailwindcss": "^3.1.0",
     "tsup": "^5.12.9",
