+++ packages/webpack-plugin/src/types/index.ts
@@ -1,22 +1,6 @@
-import type { Result } from 'postcss';
 import type webpack from 'webpack';
-import type { Config as TailwindConfig } from 'tailwindcss';
 import type { DevtoolsServer } from '../devtools/server';
-
-export type ChangedContent = { content: string; extension: string }[];
-
-export type Service = {
-  tailwindConfig: TailwindConfig | null;
-  context: any;
-  configDependencies: Set<string>;
-  transformCSS: (source: string) => Promise<Result>;
-  updateChangedContent: (changed: ChangedContent) => void;
-  clearChangedContent: () => void;
-  getCompletions: () => string;
-  ensureInit: () => void;
-  invalidateCssModule: () => void;
-  refresh: () => void;
-};
+import type { Service } from '../util';
 
 export type Compiler = {
   $tailwind: {
@@ -25,14 +9,3 @@ export type Compiler = {
     service: Service;
   };
 } & webpack.Compiler;
-
-export type DevtoolOptions = {
-  port?: number;
-  host?: string;
-};
-
-export type UserOptions = {
-  config?: TailwindConfig | string;
-  devtools?: DevtoolOptions;
-  entry?: string;
-};
