+++ packages/webpack-plugin/src/types/type.d.ts
@@ -1,0 +1,1 @@
+declare module 'tailwind-config-viewer/server';
